+++ packages/engine_flutter/lib/engine_flutter.dart
@@ -21,6 +21,7 @@ export 'src/logic/scene.dart';
 export 'src/rendering/sprite_atlas.dart';
 export 'src/rendering/animation_system.dart';
 export 'src/rendering/animation_transition_system.dart';
+export 'src/rendering/light_flicker_system.dart';
 export 'src/ui/button_menu_scene.dart';
 export 'src/ui/menu_button_atlas.dart';
 export 'src/ui/remap_menu_scene.dart';